Refactor `nanorc2h` pipeline and integrate compact generated nanorc regex highlighting in `culo`
Agent-Logs-Url: https://github.com/rofl0r/mazu-editor/sessions/c34b8f0e-a645-4dfc-bbc1-97ca7cc8c09c

+++ nanorc2h.c
@@ -380,7 +380,7 @@ static void parse_color_line(parse_result_t *pr, char *line)
 		const char *start_pos = strstr(p, "start=");
 		const char *end_pos = strstr(p, "end=");
 		if (start_pos && end_pos && end_pos > start_pos) {
-			size_t seg_len = (size_t) (end_pos - start_pos);
+			size_t seg_len = end_pos - start_pos;
 			char *segment = malloc(seg_len + 1);
 			if (segment) {
 				memcpy(segment, start_pos, seg_len);
